requirements.txt
keyboard==0.13.5
pywin32==306
pyautogui==0.9.54
pillow==10.2.0
numpy==1.24.3
opencv-python
PyQt5==5.15.11